renamespace
deprecate methods attached to [Gitlab] in preperation for moving them to client
