+++ lib/gitlab.rb
@@ -12,38 +12,22 @@ require 'gitlab/api'
 require 'gitlab/client'
 
 module Gitlab
-  extend Configuration
-
-  # Alias for Gitlab::Client.new
-  #
-  # @return [Gitlab::Client]
-  def self.client(options = {})
-    Gitlab::Client.new(options)
-  end
-
   # Delegate to Gitlab::Client
+  # @deprecated To be removed from [Gitlab] and accessed only from [Gitlab::Client]
   def self.method_missing(method, *args, &block)
-    return super unless client.respond_to?(method)
-
-    client.send(method, *args, &block)
+    if Gitlab::Client.respond_to?(method)
+      Gitlab::Client.send(method, *args, &block)
+    elsif Gitlab::Client.client.respond_to?(method)
+      Gitlab::Client.client.send(method, *args, &block)
+    else
+      super
+    end
   end
 
   # Delegate to Gitlab::Client
+  # @deprecated To be removed from [Gitlab] and accessed only from [Gitlab::Client]
   def self.respond_to_missing?(method_name, include_private = false)
-    client.respond_to?(method_name) || super
-  end
-
-  # Delegate to HTTParty.http_proxy
-  def self.http_proxy(address = nil, port = nil, username = nil, password = nil)
-    Gitlab::Request.http_proxy(address, port, username, password)
-  end
-
-  # Returns an unsorted array of available client methods.
-  #
-  # @return [Array<Symbol>]
-  def self.actions
-    hidden =
-      /endpoint|private_token|auth_token|user_agent|sudo|get|post|put|\Adelete\z|validate\z|request_defaults|httparty/
-    (Gitlab::Client.instance_methods - Object.methods).reject { |e| e[hidden] }
+    Gitlab::Client.respond_to?(method_name, include_private) ||
+      Gitlab::Client.client.respond_to?(method_name, include_private) || super
   end
 end
